server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
apscheduler==3.10.4
websockets==13.1
python-dotenv==1.0.1
